fix RegisterTouchWindow signature (ulong in C++ is not same as unlong in C#)

+++ MacBridge/WinApi/User32.cs
@@ -476,7 +476,7 @@ namespace WinApi
             return false;
         }
 
-        public static bool RegisterTouchWindow(System.IntPtr hWnd, ulong ulFlags)
+        public static bool RegisterTouchWindow(System.IntPtr hWnd, int ulFlags)
         {
 			NotImplemented(MethodBase.GetCurrentMethod());
             return false;
